集成mybatis官方代码生成器

+++ train-member/src/main/java/com/spongzi/train/member/controller/MemberController.java
@@ -20,7 +20,7 @@ public class MemberController {
     private MemberService memberService;
 
     @GetMapping("/count")
-    public Integer hello() {
+    public Long hello() {
         return memberService.count();
     }
 }
